Rename local c headers to avoid clash
HsNet.h and HsNetDef.h are exported by the `network` package as well;
this is an intermediate refactoring to spot regressions before reusing
`network`'s headers directly.

While at it, also be more explicit by using a qualified import to see
make it clearer which symbols are local to `network-bsd` and which are
reexported from `network`.