Make the high-level composite `get*Entries` operations atomic
The 'get*Entries` operations are composite and stateful and there's a
risk that invoking these from different threads might interfere with
each other via some globally shared state.

So now instead of only aquiring the lock for the individual primitive
ops we hold the lock for the duration of the complete transaction.

This also fixes a dead-lock in `getProtocolEntries` due to nested
`withLock`s.